RD-4149 Enable widgets' README.md files lazy loading (#1966)
